[WIP] Configure parameters and datasets
Add Dates, OrderedCollections.OrderedDict as dict, BusinessDays, YAML as
yml, and DateFrames and import in Kimai.jl.
Include file config.jl with functions to configure Kimai with key
parameters of the workload and sources of the datasets. Prepare the
inclusion of recovered previous sessions. Add sample data/config.yaml
and a sessions/default.yaml with data from the previous session.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>