Overhauls stats/calculation of worktime, updates/fixes
Add start and stop time for current Kimai period in "stats" of the data
dict. Entries start/stop are defined in load. Overhaul stats entries.
Fix function worktime with the calculation of worktime periods in units
defined by params from a millisecond value. Makes functions workhours
and workdays obsolete.
Add function countholidays equivalent to countbdays.
Use functions for conversion of hours and milliseconds rather than
conversion factors.
In function load_kimai, delete all data lines processed in previous
sessions. Adjust load_offdays to process integer input. If the off-day
is defined by an integer, a one-line DataFrame is return with the
integer in the count column, the start/stop date are those of the kimai
period, and the reason is the off-day type. Data to the dict is now add
in place, but return as well.
In params["Restore"], "balance" is now saved as Float of milliseconds
rather than a dict with an entry for the float and one for a
CompoundPeriod.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>