Fix Kimai dates
In load, correct unrestricted start and stop date from Date to DateTime.
Rework function load_kimai to loop over each dataframe line rather than
using vectorised operations for date conversions. The old behaviour
could fail for leap year.
Delete the date column at the end as the date is included in the
datetime of the in and out column.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>