Update balance calculation
Initialise params["tmp"] already in configure and save "calender" here
rather than in Settings to avoid saving the calendar to yaml files.
Update function calculate! to new offday structure. Use views on
vacation and sick days restricted to the current Kimai period to
calculate the number of off-days. Rename timelog to data to account for
the broader access of parameters in data. Use the calendar entry in
params["tmp"] rather than a local variable for the calendar.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>