Overhaul kwargs to manipulate balances
Move the recover section to the end of function configure and and checks
for kwargs to add or substract times to previous balances in function
correct_dates!. Use kwargs restore_vacation/sickdays instead of
vacation/sickdays_balance to define the previous balances.
Now, "balance", "restore_vacation", and "restore_sickdays" set a balance
instead of correcting it by a value. This is now done by
"correct_balance", "correct_vacation", and "correct_sickdays".
Overhaul function check_dictentry!; in argument collection of
check_dictentry!, now the whole data dict is passed and a new argument
section that defines the top level dict entry is used, while entry
defines the label of the nested dict. With the new kwarg section_in_kw,
it will be search for names in kwargs with the section name and an
underscore prepended to the kwarg name.
Use "final year" instead of "finalyear" in the config.yaml, and
final_year as kwarg.
Overhaul function load_offdays! using new rules for kwargs. Correct all
column names previously used as count to days.
Consider the previous balance in the balance entry of the tmp dict in
function add_vacationcounter!.
Add new function mstoday to convert a Real of milliseconds to an Int of
days. The number is always rounded to full days respecting rounding
rules.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>