Refactor abscence calculation
Define new calendars for sick leave, vacation, and halfdays and load the
respective leave days to define abscence by the isholiday function of
the BusinessDays package.
Add a Bool halfday to the Recover section of the config file to allow
transferring halfdays between session. All days are counted as integers,
but if half a vacation is left in a Kimai period, the halfday flag will
be set to true to preserve it.
Add helper functions in stats.jl to analyse work and abscence days.
Refactor the loading of abscence days. Currently, the calculation of
remaining vacation days, resetting vacation and balance checks do not
yet work.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>